basic ui: Properly handle escaped characters when parsing quantities
Quantity values can sometimes contain scientific notation. For example
7E+1 °F (70 degrees Fahrenheit). The Basic UI backend URL encodes this
value before rendering into the data attribute of the setpoint widget,
so the value of the raw "data-value" attribute is "7E%2B1+%C2%B0F".

To calculate the up/down steps, we of course have to parse this
value. The current logic hands the URI-encoded blob directly to
parseFloat, which (due to the escaping) fails to identify the
scientific notation. In the example above, the 70 degrees gets parsed
as 7 degrees--obviously an unintended result. This fixes the issue by
decoding the escaped characters before attempting to parse the number.